backend/requirements.txt
django
djangorestframework
python-dotenv
requests
google-cloud-vision
google-cloud-aiplatform
openai
pillow
scikit-learn